% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Custom Forest Plot}
\usage{
forest_plot(
  meta_out,
  save = FALSE,
  altplot = FALSE,
  dpi = 96,
  plot_top = 4,
  title_line = 2,
  width = 12,
  height = 11,
  title_text = "My Title",
  filename = "forest.png"
)
}
\arguments{
\item{meta_out}{Output of metafor::rma}

\item{save}{Logical}

\item{altplot}{Alternative plot}

\item{dpi}{Dots per inch}

\item{plot_top}{}

\item{title_line}{}

\item{width}{}

\item{height}{}
}
\description{
Custom Forest Plot
}
